Swagger + 接口优化

+++ src/main/java/top/fanzhengke/librarysystemspringboot/service/impl/CollegeServiceImpl.java
@@ -3,7 +3,9 @@ package top.fanzhengke.librarysystemspringboot.service.impl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import top.fanzhengke.librarysystemspringboot.domain.College;
+import top.fanzhengke.librarysystemspringboot.domain.Marjor;
 import top.fanzhengke.librarysystemspringboot.mapper.CollegeMapper;
+import top.fanzhengke.librarysystemspringboot.mapper.MarjorMapper;
 import top.fanzhengke.librarysystemspringboot.service.CollegeService;
 
 import javax.annotation.Resource;
@@ -13,6 +15,8 @@ import java.util.List;
 public class CollegeServiceImpl implements CollegeService {
     @Resource
     private CollegeMapper collegeMapper;
+    @Resource
+    private MarjorMapper marjorMapper;
 
     @Override
     public List<College> findAll() {
@@ -31,11 +35,23 @@ public class CollegeServiceImpl implements CollegeService {
 
     @Override
     public Boolean update(College college) {
+        College collegeMapperById = collegeMapper.findById(college.getId());
+        if (collegeMapperById == null) {
+            throw new RuntimeException("编辑学院失败,请检查学院ID是否正确");
+        }
         return collegeMapper.updateCollege(college);
     }
 
     @Override
     public Boolean delete(Integer id) {
+        College college = collegeMapper.findById(id);
+        if (college == null) {
+            throw new RuntimeException("删除学院失败,请检查学院ID是否正确");
+        }
+        List<Marjor> marjors = marjorMapper.findByCid(id);
+        if (marjors.size() > 0) {
+            throw new RuntimeException("请先删除旗下专业");
+        }
         return collegeMapper.delete(id);
     }
 }
